Implemented entry_state_handlers in conversationhandler.py
These Handlers will be executed when entering a state. The syntax for users is the same as for states.

The PendingState class got a few more init parameters (and an init implementation), and now resolves the states as soon as they are available in the new method await_new_state. The resolve method now maps non-blocking entry-points with an exception to an END state. Previously this special case was handled in check_update, and was not being mapped.

 Interesting changes on Class ConversationHandler:

 Method check_update:
 We no longer resolve PendingState, since it does it by itself, so the relevant code was removed, and we now directly  handle the WAITING state instead.

 Method handle_update:
 We now first map a child_conversation to a possible parent (and set the child conversation in an END state), and afterward set the timeout jobs. This ensures, we don't start unnecessary timeout jobs for ended children, and also ensures we don't enter state_entry_handlers for a false timeout.

 Method _update_state:
 We now have a few more parameters, for calling _execute_state_entry_handlers, and creating PendingState objects.

 New Method _execute_state_entry_handlers:
 has the logik for calling the state_entry_handlers

There are currently no Tests for this feature, they must be written later.