Implemented the abstract class ConversationHandlerKey with the default implementation ExampleConversationHandlerKey.
These two classes are replacing the method _get_key, and allow users to subclass them for own logik in the key creation.
Added the parameter key_builder to the __init__ method of ConversationHandler, which enables users to insert their own subclass. If none is given, the ExampleConversationHandlerKey will be used, together with the parameters per_chat, per_user and per_message. The key itself (retrieved through from_update method) is still a tuple, to assure backward compatibility.

It must be decided what to do if the key creation class returns NotImplemented, added a todo comment for this, so it will not be forgotten.