Implemented Persistence for timeout jobs
For this the class ConversationData was implemented, and integrated in existing logik, to have a clean state dict everywhere. This class holds the data relevant for restoring the timeout jobs, as well as helper methods for the persistence classes. This class is being exposed to the user in persistence classes.

This Commit includes breaking changes:
The interface of method update_conversation of BasePersistence was changed to accept a ConversationData object, and the parameters key and new_state got removed.
This forces Users to change code (when they are using persistence).

Pickle- and dict- persistence were updated to implement the new interface, but should be able to accept persistence data in the old form. THIS IS NOT WELL TESTED.

Updated existing tests to represent real test cases for this change.
The new feature itself has no tests till now.

Also changed the Type ConversationDict to represent the new change.