updated a few tests and allow state_entry_handlers to return a new state
updated tests to represent the changes.
By writing the test for state_entry_handlers, I got the idea that it would in fact be a good idea to allow state_entry_handlers to return a new state, which will be entered. for this  the timeout job creation was moved to _update state method. This also means, timeout jobs now only start after the handler is finished (also for non-blocking handlers). So the method _schedule_job_delayed got removed, since now there will be no pending states triggering on a timeout job.

Also added a few docstrings.

Fixed todo regarding keys returning NotImplemented, now we leave the handler, meaning we are not handling the corresponding update.

Tests for the persistence of timeout jobs are still missing.