Made the conversation key, as well as the conversation timeout changeable.
A User can change the Conversation key in the ConversationHandler over the context.conversation_data object.

Additionally, by setting context.conversation_data.timeout it is possible to set custom timeouts in each handler.

By changing the conversation key it should be possible to establish a conversation between different chats, as well as have the start message of a callback query conversation handler in the same handler as the other states.

maybe some warnings must be reworked by this feature