Implemented ConversationStates, and split up ConversationHandler.
Implemented the class ConversationStates, which is a datastore for the different states parameter of method __init__ of class ConversationHandler.
The parameters per_user, per_chat and per_message of method __init__ of the class ConversationHandler are now part of class DefaultConversationHandlerKey, since these parameter may be not needed on own Key implementations.
Moved ConversationHandler to _conversationhandler/conversationhandler, and split the file up, with the files conversationhandlerkey.py, defaultconversationhandlerkey.py, pendingstate.py, conversationstates.py and conversationdata.py. Each file holds the class of the same Name.

Edited __init__.py to make the new classes available for end users.

After this commit Users must Use class ConversationStates for creating an ConversationHandler, as well as class DefaultConversationHandlerKey (or another Implementation of ConversationHandlerKey)
 if they want to use different per settings.