Replaced Mocks on some locations
Thought about Mocking and future new features (which may change the integration between the mock and the tested object), and decided to replace Mock's at some locations, to prevent the tests to not detect such cases. (e.g. someone edited object and tests of the object, but since we mocked it, the changes won't be detectable).